command palette: address navigation review findings
