+++ macos/ATCTests/AppModelRuntimeTests.swift
@@ -3,138 +3,6 @@ import Testing
 import ATCAPI
 @testable import ATC
 
-/// A client whose calls can be failed on demand and/or delayed — drives
-/// reachability transitions and edit-during-refresh races.
-private nonisolated final class ScriptableClient: ATCClient, @unchecked Sendable {
-    private let inner = MockATCClient()
-    private let lock = NSLock()
-    private var _shouldFail = false
-    private var _delay: Duration?
-
-    var shouldFail: Bool {
-        get { lock.withLock { _shouldFail } }
-        set { lock.withLock { _shouldFail = newValue } }
-    }
-
-    var delay: Duration? {
-        get { lock.withLock { _delay } }
-        set { lock.withLock { _delay = newValue } }
-    }
-
-    private func gate() async throws {
-        if let delay {
-            try? await Task.sleep(for: delay)
-        }
-        if shouldFail {
-            throw ATCError.api(code: "unreachable", message: "scripted failure", sessionID: nil)
-        }
-    }
-
-    func health() async throws -> Health { try await gate(); return try await inner.health() }
-    func version() async throws -> Version { try await gate(); return try await inner.version() }
-    func sessions(includeArchived: Bool, status: SessionStatus?) async throws -> [Session] {
-        try await gate()
-        return try await inner.sessions(includeArchived: includeArchived, status: status)
-    }
-    func session(id: String) async throws -> SessionDetail {
-        try await gate(); return try await inner.session(id: id)
-    }
-    func startSession(_ request: StartSessionRequest) async throws -> SessionDetail {
-        try await gate(); return try await inner.startSession(request)
-    }
-    func terminateSession(id: String) async throws -> SessionDetail {
-        try await gate(); return try await inner.terminateSession(id: id)
-    }
-    func archiveSession(id: String) async throws -> SessionDetail {
-        try await gate(); return try await inner.archiveSession(id: id)
-    }
-    func unarchiveSession(id: String) async throws -> SessionDetail {
-        try await gate(); return try await inner.unarchiveSession(id: id)
-    }
-    func deleteSession(id: String) async throws {
-        try await gate(); try await inner.deleteSession(id: id)
-    }
-    func sendText(sessionID: String, text: String) async throws { try await gate() }
-    func sendKey(sessionID: String, key: String) async throws { try await gate() }
-    func actions() async throws -> [ATCAction] { try await gate(); return try await inner.actions() }
-    func action(name: String) async throws -> ATCAction {
-        try await gate(); return try await inner.action(name: name)
-    }
-    func createAction(_ request: ActionWriteRequest) async throws -> ATCAction {
-        try await gate(); return try await inner.createAction(request)
-    }
-    func updateAction(name: String, _ request: ActionWriteRequest) async throws -> ATCAction {
-        try await gate(); return try await inner.updateAction(name: name, request)
-    }
-    func setActionEnabled(name: String, enabled: Bool) async throws -> ATCAction {
-        try await gate(); return try await inner.setActionEnabled(name: name, enabled: enabled)
-    }
-    func deleteAction(name: String) async throws {
-        try await gate(); try await inner.deleteAction(name: name)
-    }
-    func environments() async throws -> [ATCEnvironment] {
-        try await gate(); return try await inner.environments()
-    }
-    func listDirectory(path: String, showHidden: Bool) async throws -> DirectoryListing {
-        try await gate(); return try await inner.listDirectory(path: path, showHidden: showHidden)
-    }
-    func projects(includeArchived: Bool) async throws -> [Project] {
-        try await gate(); return try await inner.projects(includeArchived: includeArchived)
-    }
-    func project(id: String) async throws -> Project { try await gate(); return try await inner.project(id: id) }
-    func createProject(name: String, workingDir: String) async throws -> Project {
-        try await gate(); return try await inner.createProject(name: name, workingDir: workingDir)
-    }
-    func renameProject(id: String, name: String) async throws -> Project {
-        try await gate(); return try await inner.renameProject(id: id, name: name)
-    }
-    func archiveProject(id: String) async throws -> Project {
-        try await gate(); return try await inner.archiveProject(id: id)
-    }
-    func unarchiveProject(id: String) async throws -> Project {
-        try await gate(); return try await inner.unarchiveProject(id: id)
-    }
-    func projectSessions(projectID: String, includeArchived: Bool, status: SessionStatus?) async throws -> [Session] {
-        try await gate()
-        return try await inner.projectSessions(
-            projectID: projectID, includeArchived: includeArchived, status: status
-        )
-    }
-    func deleteProject(id: String) async throws {
-        try await gate(); try await inner.deleteProject(id: id)
-    }
-    func workspaces(projectID: String?, includeArchived: Bool) async throws -> [Workspace] {
-        try await gate()
-        return try await inner.workspaces(projectID: projectID, includeArchived: includeArchived)
-    }
-    func workspace(id: String) async throws -> Workspace {
-        try await gate(); return try await inner.workspace(id: id)
-    }
-    func createWorkspace(projectID: String, name: String) async throws -> Workspace {
-        try await gate(); return try await inner.createWorkspace(projectID: projectID, name: name)
-    }
-    func renameWorkspace(id: String, name: String) async throws -> Workspace {
-        try await gate(); return try await inner.renameWorkspace(id: id, name: name)
-    }
-    func archiveWorkspace(id: String) async throws -> Workspace {
-        try await gate(); return try await inner.archiveWorkspace(id: id)
-    }
-    func unarchiveWorkspace(id: String) async throws -> Workspace {
-        try await gate(); return try await inner.unarchiveWorkspace(id: id)
-    }
-    func deleteWorkspace(id: String) async throws {
-        try await gate(); try await inner.deleteWorkspace(id: id)
-    }
-    func workspaceSessions(workspaceID: String, includeArchived: Bool, status: SessionStatus?) async throws -> [Session] {
-        try await gate()
-        return try await inner.workspaceSessions(
-            workspaceID: workspaceID, includeArchived: includeArchived, status: status
-        )
-    }
-    func attachURL(sessionID: String) -> URL { inner.attachURL(sessionID: sessionID) }
-    func attachHeaders() -> [String: String] { inner.attachHeaders() }
-}
-
 /// Runtime lifecycle: add/edit/delete touch only the affected runtime,
 /// selection cleanup, reachability transitions, stale-result harmlessness.
 @MainActor
